Nuevo método en la Fábrica de Enemigos
La fábrica de enemigos ahora posee un método
para crear un enemigo correspondiente con
el enum recibido.
El ControladorEnemigo ahora usa este método
en vez de crearlo el mismo.

+++ Proyecto/core/src/edu/uns/galaxian/entidades/autonoma/enemigo/Enemigo.java
@@ -24,7 +24,7 @@ public class Enemigo extends EntidadViva implements Autonomo  {
 	private ColisionadorEnemigo colisionador;
 	private InteligenciaArtificial inteligencia;
 	
-	public Enemigo(int xPos, int yPos, NaveEnemigo nave, ControladorEnemigo controlador){
+	public Enemigo(float xPos, float yPos, NaveEnemigo nave, ControladorEnemigo controlador){
 		super(new Vector2(xPos, yPos), nave.getRotacionInicial(), nave.getVidaMax());
 		this.nave = nave;
 		this.controladorEnemigo = controlador;
